requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-rest-auth==0.9.5
flake8==3.7.9
ipython==7.13.0
ipdb==0.13.2
